Several fixes and improvements (that I couldn't commit separately)
- Fixes Pairwise Euclidean distance computation. Since initially everything
was done with the radial basis function, the pairwise Euclidean calculation
returned the squared of the pairwise distance.

- Fixes order of evaluations for the gradient/Normal. Those evaluation must
occur before updating the Covariance matrix with actual covariance value
after applying the covariance function.

- The create() function now must be called create<true/false>() to tell
the normal at the training data should be estimated as well. The choice for
a templated function makes possible the compiler and optimizer ot get rid
of unused branches due to if/else statements at compilation time.

- Adds other signatures of the evaluate function to focus on what's
requested, and not loose time computing useless information when the
function is being called.

- The computeTangentBasis() seems to be bad as before, this is easier.
However, this is not general, since if N = UnitX will make Tx to be UnitX,
and in consequence Ty = N.cross(Tx) = 0. But what are the odds?

- Comment cleaning and update of description of the Model struct.