Generate charts with the size equals to the variance at its center
This value can be very small depending of scaling, but this seems like a
good place to store this value that will be most likely using for the
planning.

It also updates the test_gaussian.cpp to avoid compilation errors. It is
also consistent with the updates of the regressor class.