Converts indentations to spaces and make some C++ style consistent
@memmone,

To me, this is important when sharing and collaborating at code level.
Though I read better my style, I don't mind different styles in the same
pack or library.

But what's really hard to read is different styles in the same file, so
my way to understand other's code is formatting as close to my style as
possible, or fixing the own file style using the majority of cases, like
bracing.

So, the thing is, I don't think I changed anything here, only formatting,
so ignore or delete this commit if undesired.