Adds support for uncertainty in the training points
The uncertainty or noise does not have to be the same for all. Each
training point can have different value. This is what makes Gaussian
Processes ideal for sensor fusion, because each point can have its
associated uncertainty.

For point clouds from the kinect, a value of sigma squared = 0.000001 gives
good results, which is a resolution of 0.001m -> 1mm.

cc: @Tabjones