Adds Choleski-based computation
I didn't believed it, but it is true. No need to .inverse(), Choleski is
much more efficient.

For the test obj.pcd, the creation was reduced from 1s to 0.2s, and for the
fakeDeterministicSampling, from 30min to 7min.

Also reduces the mem size of the model, there values that were useless in
further computations.