Adds the publishing of the estimated blob as a point cloud
This is useful for planning to be close to the surface. This point cloud
is configured to be the input of the moveit octomap server to create a
collision map of the estimated shape to avoid crashing with it.

Aware of the following changes

* What used to be published in the /estimated_model is now published in the
topic /training_data, still as a point cloud
* The topic /estimated_model now contains the explicit shape representation
of the shape (to differentiate from the implicit i.e. the GP) already
mapped to the real world

@Tabjones cc'ing due to the change in what's being published in the topics.